requirements.txt
openai
python-dotenv
chromadb
ollama
requests
typing-extensions